refactor(python-sdk): mirror the e2b.api layout for envd RPC clients
Shared plumbing (JSON codec, proxy narrowing, pool tuning) moves to
e2b/envd/__init__.py; the flavor-specific transports and client factories
move to e2b/envd/client_sync/ and e2b/envd/client_async/, each exposing
get_transport/create_rpc_client/as_stream under the same names like
e2b.api.client_sync and client_async do. Interceptors stay in
e2b/envd/interceptors.py; the transport.py and client.py modules are gone.
