+++ packages/python-sdk/e2b/sandbox_sync/commands/pty.py
@@ -24,7 +24,7 @@ from e2b.envd.rpc import (
     handle_rpc_exception_with_health,
     timeout_to_ms,
 )
-from e2b.envd.client import as_stream, create_rpc_client
+from e2b.envd.client_sync import as_stream, create_rpc_client
 from e2b.sandbox.commands.command_handle import PtySize
 from e2b.sandbox_sync.commands.command_handle import CommandHandle
